+++ src/components/user/contact/ContactForm.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import styles from './Contact.module.css';
+
+export default function ContactForm(){
+    return(
+        <div className={styles.all}>
+            <div className={styles.title}><p>sdffsd</p></div>
+        </div>
+    );
+}
