+++ src/pages/_app.tsx
@@ -3,8 +3,8 @@ import { useRouter } from 'next/router';
 
 import type { AppProps } from 'next/app';
 
-import Header from '@/components/user/Header';
-import Footer from '@/components/user/Footer';
+import Header from '@/components/user/common/Header';
+import Footer from '@/components/user/common/Footer';
 
 export default function App({ Component, pageProps }: AppProps) {
   const router = useRouter();
