+++ src/pages/contact/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import ContactForm from '@/components/user/contact/ContactForm';
+
+export default function ContactPage(){
+    return(
+        <ContactForm />
+    );
+}
