--- src/pages/login.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function Login() {
-  return (
-    <div>
-      <div></div>
-    </div>
-  );
-}
