+++ src/pages/login/index.tsx
@@ -1,0 +1,7 @@
+import LoginForm from '@/components/user/login/LoginForm';
+
+export default function LoginPage() {
+  return (
+    <LoginForm />
+  );
+}
