+++ src/pages/mypage/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import styles from './Mypage.module.css';
-import MypageNav from '@/components/user/MypageNav';
+import MypageNav from '@/components/user/mypage/MypageNav';
 
 interface Reservation {
   id: number;
